requirements.txt
spotipy==2.24.0
python-dotenv==1.0.1
click==8.1.7